[#7050] Avoid compiler bug
