+++ grpc/src/main/java/com/navercorp/pinpoint/grpc/client/SocketIdClientInterceptor.java
@@ -21,7 +21,6 @@ import io.grpc.CallOptions;
 import io.grpc.Channel;
 import io.grpc.ClientCall;
 import io.grpc.ClientInterceptor;
-import io.grpc.ForwardingClientCall.SimpleForwardingClientCall;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 
@@ -37,17 +36,17 @@ public class SocketIdClientInterceptor implements ClientInterceptor {
     @Override
     public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(MethodDescriptor<ReqT, RespT> method, CallOptions callOptions, Channel next) {
         final ClientCall<ReqT, RespT> clientCall = next.newCall(method, callOptions);
-        final ClientCall<ReqT, RespT> forwardingClientCall = new SimpleForwardingClientCall<ReqT, RespT>(clientCall) {
+        final ClientCall<ReqT, RespT> forwardClientCall = new ForwardClientCall<ReqT, RespT>(clientCall) {
             @Override
             public void start(Listener<RespT> responseListener, Metadata headers) {
                 final String socketId = nextSocketId();
                 headers.put(Header.SOCKET_ID, socketId);
                 super.start(responseListener, headers);
             }
-
         };
-        return forwardingClientCall;
+        return forwardClientCall;
     }
+
     private String nextSocketId() {
         return String.valueOf(socketIdAllocator.incrementAndGet());
     }
