+++ grpc/src/main/java/com/navercorp/pinpoint/grpc/client/SupportCommandCodeClientInterceptor.java
@@ -23,7 +23,6 @@ import io.grpc.CallOptions;
 import io.grpc.Channel;
 import io.grpc.ClientCall;
 import io.grpc.ClientInterceptor;
-import io.grpc.ForwardingClientCall;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 
@@ -41,7 +40,7 @@ public class SupportCommandCodeClientInterceptor implements ClientInterceptor {
     @Override
     public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(MethodDescriptor<ReqT, RespT> method, CallOptions callOptions, Channel next) {
         final ClientCall<ReqT, RespT> clientCall = next.newCall(method, callOptions);
-        final ClientCall<ReqT, RespT> forwardingClientCall = new ForwardingClientCall.SimpleForwardingClientCall<ReqT, RespT>(clientCall) {
+        final ClientCall<ReqT, RespT> forwardClientCall = new ForwardClientCall<ReqT, RespT>(clientCall) {
             @Override
             public void start(Listener<RespT> responseListener, Metadata headers) {
 
@@ -63,7 +62,7 @@ public class SupportCommandCodeClientInterceptor implements ClientInterceptor {
             }
 
         };
-        return forwardingClientCall;
+        return forwardClientCall;
     }
 
 }
