+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/provider/grpc/SpanGrpcDataSenderProvider.java
@@ -32,9 +32,9 @@ import com.navercorp.pinpoint.grpc.client.HeaderFactory;
 import com.navercorp.pinpoint.profiler.context.module.SpanConverter;
 import com.navercorp.pinpoint.profiler.context.thrift.MessageConverter;
 import com.navercorp.pinpoint.profiler.sender.DataSender;
-import com.navercorp.pinpoint.profiler.sender.grpc.DiscardClientInterceptor;
-import com.navercorp.pinpoint.profiler.sender.grpc.DiscardEventListener;
-import com.navercorp.pinpoint.profiler.sender.grpc.LoggingDiscardEventListener;
+import com.navercorp.pinpoint.grpc.client.interceptor.DiscardClientInterceptor;
+import com.navercorp.pinpoint.grpc.client.interceptor.DiscardEventListener;
+import com.navercorp.pinpoint.grpc.client.interceptor.LoggingDiscardEventListener;
 import com.navercorp.pinpoint.profiler.sender.grpc.ReconnectExecutor;
 import com.navercorp.pinpoint.profiler.sender.grpc.SpanGrpcDataSender;
 import io.grpc.ClientInterceptor;
